pyproject.toml
[project]
name = "langchain-rss-news"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "chromadb>=1.3.5",
    "feedparser>=6.0.12",
    "langchain>=1.0.8",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
